Tests, etc: refactor in light of addition of Journal model (#19)
Now that we have a `Journal` model (which "owns" `Account` models) we
should update our testing helpers to account for the journal instances
more explicitly. This PR does just that.

Also we add the notion of a 'current period' to `Journal` and remove the
notion of a globally configured "accounting period": a `Journal` now
controls what is the current accounting period in any given context for
objects which descend from it (accounts, line items, etc.). Later work
can further modify existing scopes to require that journal object to use
as a fallback when a more specific period is not needed (and to ensure
all e.g. ledger line items are always scoped to a specific journal).

Related to #16. Closes #19.

-----

Automated Test Results:

OK (191 tests, 319 assertions)